+++ woocommerce-gateway-ebanx/woocommerce-gateway-ebanx.php
@@ -5,7 +5,7 @@
  * Description: Offer Latin American local payment methods & increase your conversion rates with the solution used by AliExpress, AirBnB and Spotify in Brazil.
  * Author: EBANX
  * Author URI: https://www.ebanx.com/business/en
- * Version: 1.1.0
+ * Version: 1.1.1
  * License: MIT
  * Text Domain: woocommerce-gateway-ebanx
  * Domain Path: /languages
@@ -35,7 +35,7 @@ if (!class_exists('WC_EBANX')) {
          *
          * @var string
          */
-        const VERSION = '1.1.0';
+        const VERSION = '1.1.1';
 
         const DIR = __FILE__;